refactor(_cadqueryapi): extract _curves / _display / _io submodules
Step 2 of the package split. Pulls ~700 lines out of _core.py:

* _curves.py (333)  — Bezier / B-spline / circle / ellipse constructors
* _display.py (165) — tessellate, collect_*, DefaultDisplayOptions, show_cad
* _io.py (339)      — CADFileType, save_cad / save_as_STP / import_cad,
                      change_placement (uses _CQPlacement only as type hint)

deserialise_shape uses a local import of the curve constructors to avoid a
module-level _core ↔ _curves cycle. Public API surface unchanged
